refactor: extract shared _SqliteTracker base class
Three trackers (budget, usage, relevance) had an identical 8-line init
block (parent mkdir, threading.Lock, sqlite3 connect with
check_same_thread=False, WAL pragma, executescript) plus identical
close/__del__ methods.

Hoist that machinery into _SqliteTracker. Subclasses declare _SCHEMA
as a class attribute and inherit __init__/close/__del__. RelevanceTracker
keeps its own __init__ to receive alpha/decay/epsilon but delegates
connection setup to super().

Net: -21 lines, no behavior change, 416 tests pass unchanged.
